# backend/openapi.yaml
components:
  schemas:
...
    AuditEntry:
      description: |-
        One row of the audit journal: who (`actor`) did what (`event_type` on
        `entity_type`/`entity_id`) when (`occurred_at`), with the event `payload`.

        `series_id` is the tenant dimension prepared for per-`ProjectId` tenancy
        (decision 9.2) and is `NULL` in v1. `payload` is the raw event serialized
        as JSON (generic, so any context's events fit the same row).
      properties:
        actor:
          oneOf:
          - $ref: '#/components/schemas/UserId'
          - type: 'null'
        block_id:
          oneOf:
          - $ref: '#/components/schemas/BlockId'
          - type: 'null'
        entity_id:
          type: string
        entity_type:
          type: string
        event_type:
          type: string
        id:
          format: uuid
          type: string
        occurred_at:
          format: date-time
          type: string
        payload: {}
        series_id:
          format: uuid
          type:
          - string
          - 'null'
      required:
      - id
      - entity_type
      - entity_id
      - event_type
      - payload
      - occurred_at
      type: object
    BlockId:
      description: Opaque identifier for a `Block` aggregate.
      format: uuid
      type: string
...
    EpisodeId:
      description: Opaque identifier for an `Episode` aggregate.
      format: uuid
      type: string
...
    InviteMemberRequest:
      properties:
        role:
          $ref: '#/components/schemas/Role'
          description: Proposed role for the invited user (pending until they accept).
        user_id:
          description: OIDC `sub` of the user to invite to the block.
          type: string
      required:
      - user_id
      - role
      type: object
...
    MembershipStateKind:
      description: |-
        Membership lifecycle state in the read model.

        `snake_case` serialization keeps the Postgres `state` text column stable
        and human-readable.
      enum:
      - pending
      - active
      type: string
    MembershipView:
      description: |-
        Complete membership read model row for one `(block_id, user_id)` pair.

        `joined_at` is the timestamp of the `InvitationAccepted` event, sourced
        from the event stream (not from aggregate state).
      properties:
        block_id:
          $ref: '#/components/schemas/BlockId'
        joined_at:
          format: date-time
          type: string
        role:
          $ref: '#/components/schemas/Role'
        state:
          $ref: '#/components/schemas/MembershipStateKind'
        user_id:
          $ref: '#/components/schemas/UserId'
      required:
      - block_id
      - user_id
      - role
      - state
      - joined_at
      type: object
...
    ProblemDetails:
      description: |-
        An RFC 9457 problem document (ADR-031 D1).

        Every error response (status ≥ 400) produced by the API is an instance of
        this document: `type` (derived from `code`), constant English `title`,
        `status`, stable `code`, localized `detail` (English until Tranche 3), and
        the `trace_id` extension for otel correlation.
      example:
        code: concurrency.version-mismatch
        detail: Version conflict
        extensions:
          current_version: 3
          expected_version: 2
        status: 409
        title: Version conflict
        trace_id: 4bf92f3577b34da6a3ce929d0e0e4736
        type: https://docs.breakdown.example/problems/scene.already-scheduled
      properties:
        code:
          description: Stable machine identity `{context}.{reason}` (the client contract).
          example: concurrency.version-mismatch
          type: string
        detail:
          description: Human-readable explanation; localized server-side (Tranche 3).
          example: Version conflict
          type: string
        extensions:
          additionalProperties: {}
          description: Declared S0/S1 extension fields, if any (ADR-031 D4).
          propertyNames:
            type: string
          type:
          - object
          - 'null'
        status:
          description: Canonical HTTP status of this problem.
          example: 409
          format: int32
          minimum: 0
          type: integer
        title:
          description: Constant English title (never localized; cacheable, spec-stable).
          type: string
        trace_id:
          description: W3C trace id of the request's otel span (support correlation).
          example: 4bf92f3577b34da6a3ce929d0e0e4736
          type: string
        type:
          description: Dereferencable documentation anchor, derived from the `code`.
          example: https://docs.breakdown.example/problems/scene.already-scheduled
          type: string
      required:
      - type
      - title
      - status
      - code
      - detail
      - trace_id
      type: object
    ProjectId:
      description: |-
        Opaque identifier for a `Project` — the tenant-level production container.

        `ProjectId` is an opaque UUIDv7 value type introduced by the
        `introduce-season-block-episode-hierarchy` change and renamed from
        `SeriesId` by issue #591 (ADR-035 D1/S1). The rename is **1:1**: the
        position, the UUIDv7 values and the tenant boundary are unchanged (ADR-035
        B1) — only the production-form-specific *word* is gone, so the container no
        longer claims to be a TV show.

        It remains a seam rather than an aggregate: every hierarchy entity
        (Season, Block, Episode) references it, but no `Project` aggregate exists
        yet. The children below it stay `Season`/`Block`/`Episode` until a second
        production form actually lands (ADR-035 D2/D3).

        **Wire compatibility (issue #591, layers 2 and 3 explicitly out of the
        rename).** The *type* is `#[serde(transparent)]` over `Uuid`, so a field
        typed `ProjectId` still serializes as a bare UUID string — the JSON value is
        unchanged by this rename. What deliberately keeps the old spelling is the
        **field name** on persisted payloads: event fields and read-model view
        fields carry `#[serde(rename = "series_id")]`, and the projection columns and
        OpenAPI fields keep `project_id` until the dedicated layer-3 change (a
        breaking ADR-021 `/v2` migration).
      format: uuid
      type: string
...
    Role:
      description: |-
        Block-scoped costume-department role.

        Roles are **domain-local** and **block-scoped** (Decision 4): the same
        `UserId` may hold a different `Role` in two blocks of the same season
        because staff rotate roles at Block boundaries. The initial v1 set is
        `CostumeDesigner` + `WardrobeSupervisor`, plus `CostumeAssistant` which is
        the default role assigned to the block creator during the owner bootstrap
        (see `BootstrapOwner`).

        **Ubiquitous Language is English.** The enum variants and their `snake_case`
        serde form are the canonical domain vocabulary, so events and projection
        rows are persisted as English strings (`"costume_designer"`,
        `"wardrobe_supervisor"`, `"costume_assistant"`).

        The enum is **open for additive extension** (see `block-membership` spec,
        "Initial role set"): adding a new variant is a non-breaking change for
        writers, but renaming or removing an existing variant is a breaking change
        requiring a separate proposal. Variants are serialized by their stable
        `snake_case` name, so events/rows written today stay readable after a
        future addition.

        `OpsAdmin` (issue #409) is the one deployment-scoped capability: it rides
        on block-scoped membership rows like every other role, but the ops
        predicates check it across **all** blocks (`has_active_ops_role`) because
        projector health is deployment-wide infrastructure state, not production
        data. It cannot be granted through the block-scoped membership API by
        non-ops callers (API-edge escalation guard in the `invite_member` /
        `grant_role` handlers); the first ops holder is bootstrapped via the
        `OPS_ADMIN_SUBS` environment allowlist (see the API policy).
      enum:
      - costume_designer
      - wardrobe_supervisor
      - costume_assistant
      - ops_admin
      type: string
...
    SeasonId:
      description: Opaque identifier for a `Season` aggregate.
      format: uuid
      type: string
...
    UserId:
      description: |-
        Opaque identifier for a user, wrapping the OIDC `sub` claim.

        `UserId` references the authenticated principal without ever decoding,
        storing, or dereferencing identity attributes in `core`. The backend only
        trusts the IdP-issued `sub`; account lifecycle lives exclusively in the
        OIDC provider (ADR-010). Unlike the hierarchy ids, `UserId` is *not* a
        UUIDv7 — it is the raw string subject the IdP assigns.
      type: string
...
info:
  description: |-
    Costume scheduling API for Breakdown RS (ADR-006). Routes are mounted under the /v1 path prefix; the path version is bumped only on a breaking wire change (ADR-021).
    Error contract (ADR-031): every error response (status ≥ 400) is an RFC 9457 `application/problem+json` document carrying a stable `code` (`{context}.{reason}` kebab-case) and a `trace_id` extension. The full code registry is exposed under `x-code-registry`; per-code documentation lives at `docs/errors/`.
  license:
    identifier: AGPL-3.0
    name: AGPL-3.0
  title: Breakdown RS API
  version: v1
openapi: 3.1.0
paths:
...
  /v1/blocks/{id}/audit:
    get:
      operationId: get_block_audit
      parameters:
      - description: Block id
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: episode_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/EpisodeId'
          - type: 'null'
      - in: query
        name: season_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/SeasonId'
          - type: 'null'
      - in: query
        name: series_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/ProjectId'
          - type: 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditEntry'
                type: array
          description: Audit journal entries for the block, newest first
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Missing or malformed X-Active-Block header
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Caller is not an active member of the active block (X-Active-Block header)
      tags:
      - handlers
  /v1/blocks/{id}/members:
    get:
      description: 'Gated `BlockMember`: the caller must be an active member of the block.'
      operationId: list_members
      parameters:
      - description: Block id
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          format: int64
          type:
          - integer
          - 'null'
      - in: query
        name: episode_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/EpisodeId'
          - type: 'null'
      - in: query
        name: season_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/SeasonId'
          - type: 'null'
      - in: query
        name: series_id
        required: false
        schema:
          oneOf:
          - $ref: '#/components/schemas/ProjectId'
          - type: 'null'
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MembershipView'
                type: array
          description: Members of the block (active and pending)
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Missing or malformed X-Active-Block header
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Caller is not an active member of the active block
      summary: List the members of a block (paginated).
      tags:
      - handlers
    post:
      description: |-
        Gated `BlockMember`: the caller must be an active member of the active
        block (see `authorize_middleware`). The actor is the authenticated caller.
      operationId: invite_member
      parameters:
      - description: Block id
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteMemberRequest'
        required: true
      responses:
        '204':
          description: Invitation created (pending until the invitee accepts)
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Invalid request (e.g., malformed user_id)
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Caller is not an active member of the active block (X-Active-Block header)
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Conflicting state (e.g., user is already a member)
      summary: Invite a user to the block with a proposed role (pending until accepted).
      tags:
      - handlers
...
  /v1/blocks/{id}/members/{user_id}:
    delete:
      description: |-
        Gated `BlockMember`: the caller must be an active member. The targeted
        `user_id` is taken from the path.
      operationId: remove_member
      parameters:
      - description: Block id
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: OIDC sub of the member to remove
        in: path
        name: user_id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Member removed
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Caller is not an active member of the active block
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Target user is not a member of the block
        '422':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Validation error
      summary: Remove an active member from the block.
      tags:
      - handlers
    get:
      description: 'Gated `BlockMember`: the caller must be an active member of the block.'
      operationId: get_member
      parameters:
      - description: Block id
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
      - description: OIDC sub of the member
        in: path
        name: user_id
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipView'
          description: Membership of the user in the block
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Missing or malformed X-Active-Block header
        '403':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Caller is not an active member of the active block
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Membership not found
      summary: Fetch a single membership (a block member's role and state).
      tags:
      - handlers
...